fix(ci): python-coverage-shards — keep github/bitbucket clients under file-size guardrail
The BitBucket PR-monitor integration grew github_client.py (1522) and
bitbucket_client.py (1581) past the 1500-line first-party guardrail in
test_first_party_code_files_stay_under_line_limit. Decompose by extracting
cohesive units to sibling modules (no behavior change):

- github_graphql.py: the _GQL_* query/mutation string constants (client now 1286)
- bitbucket_client_errors.py: BitBucketClientError, BitBucketAuth, reason codes,
  and _is_bitbucket_merge_in_progress_body (client now 1444)

Both are re-exported from their original modules (bitbucket_client gains an
explicit __all__ so mypy treats the re-exports as exported), so existing
imports keep working unchanged.
